requirements.txt
# Dashboard requirements
dash==2.14.2
plotly==5.17.0
redis==5.0.1
python-binance==1.0.19
pandas==2.1.4
numpy==1.24.3

# Bot requirements (for self_optimizing_bot.py)
scikit-learn==1.3.2
aiohttp==3.9.1
python-dotenv==1.0.0
requests==2.31.0
websocket-client==1.6.4